[WiP] Temporarily disable, while kinda fixing, float32 input args

+++ tools/gostd/jtypes/jtypes.go
@@ -392,7 +392,7 @@ var UIntPtr = &Info{
 }
 
 var Float32 = &Info{
-	FullName:             "Double",
+	FullName:             "Float32",
 	FullNameDoc:          "Double",
 	ArgClojureType:       "",
 	ArgFromClojureObject: "",
